requirements.txt
streamlit==1.15.2
gitpython==3.1.30
requests==2.28.2
sentence-transformers==2.2.2
chromadb==0.4.18
langchain
langchain-community==0.2.13
reportlab==3.6.12
ollama==0.4.7
torch==2.1.0
torchvision==0.16.0
